CHANGE: 'Demote' VirtualMouseInput to just sample.
