+++ Assets/Samples/GamepadMouseCursor/Unity.InputSystem.VirtualMouseInput.asmdef
@@ -1,0 +1,15 @@
+{
+    "name": "Unity.InputSystem.VirtualMouseInput",
+    "references": [
+        "GUID:75469ad4d38634e559750d17036d5f7c"
+    ],
+    "optionalUnityReferences": [],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": []
+}
