+++ Packages/com.unity.inputsystem/Documentation~/Gamepad.md
@@ -6,7 +6,7 @@ A gamepad can have additional Controls, such as a gyro, which the Device can exp
 
 Gamepad support guarantees the correct location and functioning of Controls across platforms and hardware. A PS4 DualShock controller, for example, is meant to look identical regardless of which platform it is supported on. A gamepad's south face button should always be the lowermost face button.
 
->NOTE: In case you want to use the gamepad for driving mouse input, see [`Virtual Mouse Input`](./UISupport#virtual-mouse-input).
+>NOTE: In case you want to use the gamepad for driving mouse input, there is a sample called `Gamepad Mouse Cursor` you can install from the package manager UI when selecting the Input System package. The sample demonstrates how to set up gamepad input to drive a virtual mouse cursor.
 
 ## Controls
