+++ Packages/com.unity.inputsystem/Documentation~/UISupport.md
@@ -45,11 +45,3 @@ The Input System can also handle multiple separate UI instances on the screen co
 Unlike the [`EventSystem`](https://docs.unity3d.com/Manual/script-EventSystem.html) component, you can have multiple [`MultiplayerEventSystems`](../api/UnityEngine.InputSystem.UI.MultiplayerEventSystem.html) active in the scene at the same time. That way, you can have multiple players, each with their own [`InputSystemUIInputModule`](../api/UnityEngine.InputSystem.UI.InputSystemUIInputModule.html) and [`MultiplayerEventSystem`](../api/UnityEngine.InputSystem.UI.MultiplayerEventSystem.html) components, and each player can have their own set of Actions driving their own UI instance. If you are using the [`PlayerInput`](Components.md#playerinput-component) component, you can also set up [`PlayerInput`](Components.md#playerinput-component) to automatically configure the player's [`InputSystemUIInputModule`](../api/UnityEngine.InputSystem.UI.InputSystemUIInputModule.html) to use the player's Actions. See the documentation on [`PlayerInput`](Components.md#ui-input) to learn how.
 
 The properties of the [`MultiplayerEventSystem`](../api/UnityEngine.InputSystem.UI.MultiplayerEventSystem.html) component are identical with those from the [`EventSystem`](https://docs.unity3d.com/Manual/script-EventSystem.html), as documented [here](https://docs.unity3d.com/Manual/script-EventSystem.html). Additionally, the [`MultiplayerEventSystem`](../api/UnityEngine.InputSystem.UI.MultiplayerEventSystem.html) component adds a [`Player Root`](../api/UnityEngine.InputSystem.UI.MultiplayerEventSystem.html#UnityEngine_InputSystem_UI_MultiplayerEventSystem_playerRoot) property, which you can set to a GameObject which contains all the UI [selectables](https://docs.unity3d.com/Manual/script-Selectable.html) this event system should handle in its hierarchy. Mouse input that this event system processes then ignores any UI [selectables](https://docs.unity3d.com/Manual/script-Selectable.html) which are not on any game object in the hierarchy under  [`Player Root`](../api/UnityEngine.InputSystem.UI.MultiplayerEventSystem.html#UnityEngine_InputSystem_UI_MultiplayerEventSystem_playerRoot).
-
-## Virtual Mouse Input
-
->NOTE: There is a sample called `Gamepad Mouse Cursor` you can install from the package manager UI when selecting the Input System package. The sample demonstrates how to set up gamepad input to drive a virtual mouse cursor.
-
-You may want to use the gamepad, for example, to drive a mouse cursor in the UI and perform operations such as scrolling up and down. You can achieve this using the [`VirtualMouseInput`](../api/UnityEngine.InputSystem.UI.VirtualMouseInput.html) component. Simply add a `GameObject` to represent the mouse cursor to your [`Canvas`](https://docs.unity3d.com/ScriptReference/Canvas.html), add the [`VirtualMouseInput`](../api/UnityEngine.InputSystem.UI.VirtualMouseInput.html) component to it, and configure the actions on the component appropriate. Also, to correlate the mouse position to the transform of the cursor UI object, set the "Cursor Transform" property to the respective [`RectTransform`](https://docs.unity3d.com/ScriptReference/RectTransform.html) and "Canvas" to the canvas to which the cursor should be restricted.
-
-![Virtual Mouse Input Component](./Images/VirtualMouseInput.png)
